requirements.txt
BeautifulSoup==3.2.1
Django==1.5.5
Pillow==2.3.0
PyYAML==3.10
Scrapy==0.18.4
South==0.8.2
argparse==1.2.1
cssselect==0.9.1
django-extensions==1.2.5
django-taggit==0.10
goose-extractor==1.0.6
jieba==0.31
lxml==3.2.5
nltk==2.0.4
six==1.4.1
wsgiref==0.1.2
